created orderController

+++ backend/src/main/java/com/example/springboot_mysql_demo/Services/OrderService.java
@@ -2,6 +2,7 @@ package com.example.springboot_mysql_demo.Services;
 
 
 import com.example.springboot_mysql_demo.Models.Order;
+import com.example.springboot_mysql_demo.Models.User;
 import com.example.springboot_mysql_demo.Repositories.OrderRepository;
 import com.example.springboot_mysql_demo.Repositories.OrderedProductRepository;
 import com.example.springboot_mysql_demo.Repositories.UserRepository;
@@ -24,8 +25,18 @@ public class OrderService {
     }
 
 
-    public Order createOrder (Order order){
-//      validations
+    public Order createOrder (Long userID) {
+        Order order = new Order();
+        User user = userRepo.findById(userID).
+                orElseThrow(() -> new EntityNotFoundException(
+                "User with ID: " + userID + " not found, could not create order."));
+        if (order.getBasketTotal() > 0) {
+            throw new IllegalArgumentException("Basket total cannot be a negative number.");
+        }
+        if (!userRepo.existsById(order.getUser().getId())){
+            throw new IllegalArgumentException("Could not create order as user ID not found");
+        }
+        order.setUser(user);
         return orderRepo.save(order);
     }
 
@@ -43,11 +54,18 @@ public class OrderService {
         Order existingOrder = orderRepo.findById(id).
                 orElseThrow(()->
                         new EntityNotFoundException("Order not found for id: "+id));
+        if (newOrder.getProductList() == existingOrder.getProductList()
+                && newOrder.getBasketTotal() == existingOrder.getBasketTotal()){
+            throw new IllegalArgumentException("New order is equal to existing order");
+        }
         return orderRepo.save(existingOrder);
     }
 
     public void deleteOrder (Long id){
-//        validations
+    if (!orderRepo.existsById(id)){
+        throw new EntityNotFoundException(
+                "Order with ID: " + id + " was not found and therefore could not be deleted.");
+    }
         orderRepo.deleteById(id);
     }
